Changed level info from a float to an int

+++ Balancer/GUI_Balancer.py
@@ -829,7 +829,7 @@ class HTMNetwork(QtGui.QWidget):
         # Fix this please. It must initialise the views to display.
         for i in range(len(self.HTMNetworkGrid)):
             # Add one to i since i starts at zero
-            displayLevel = math.floor((i+1)/self.numLevels)
+            displayLevel = int(math.floor((i+1)/self.numLevels))
             displayLayer = i % self.numLayers
             self.setLevel(self.HTMNetworkGrid[i], displayLevel)
             self.setLevel(self.inputGrid[i], displayLevel)
